D3ASIM-960: REfactored Plotlygrapg class to separate file.
